refactor: site data to a single file
